cli/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
GitPython>=3.1.0
Jinja2>=3.1.0
PyYAML>=6.0.0
questionary>=2.0.0
pathspec>=0.11.0